[RB TR] - consolidate manageSubscription, manageMembership and manageContribution into a new singular component manageProduct
