--- app/client/components/accountoverview/manageSubscription.tsx
@@ -1,435 +1,0 @@
-import { css } from "@emotion/core";
-import { palette, space } from "@guardian/src-foundations";
-import { headline, textSans } from "@guardian/src-foundations/typography";
-import { Link } from "@reach/router";
-import React from "react";
-import { formatDateStr } from "../../../shared/dates";
-import {
-  augmentInterval,
-  getFuturePlanIfVisible,
-  getMainPlan,
-  isGift,
-  isPaidSubscriptionPlan,
-  isSixForSix,
-  ProductDetail
-} from "../../../shared/productResponse";
-import {
-  hasDeliveryRecordsFlow,
-  ProductTypes,
-  shouldHaveHolidayStopsFlow
-} from "../../../shared/productTypes";
-import { maxWidth } from "../../styles/breakpoints";
-import { LinkButton } from "../buttons";
-import { CallCentreEmailAndNumbers } from "../callCenterEmailAndNumbers";
-import { DeliveryAddressDisplay } from "../delivery/address/deliveryAddressDisplay";
-import { FlowStartMultipleProductDetailHandler } from "../flowStartMultipleProductDetailHandler";
-import { navLinks } from "../nav";
-import { PageHeaderContainer, PageNavAndContentContainer } from "../page";
-import { CardDisplay } from "../payment/cardDisplay";
-import { DirectDebitDisplay } from "../payment/directDebitDisplay";
-import { PayPalDisplay } from "../payment/paypalDisplay";
-import { ProblemAlert } from "../ProblemAlert";
-import { ProductDescriptionListTable } from "../productDescriptionListTable";
-import { SupportTheGuardianButton } from "../supportTheGuardianButton";
-import { ErrorIcon } from "../svgs/errorIcon";
-import { GiftIcon } from "../svgs/giftIcon";
-import { RouteableStepProps } from "../wizardRouterAdapter";
-
-export const ManageSubscription = (props: RouteableStepProps) => {
-  const subHeadingTitleCss = `
-    ${headline.small()};
-    font-weight: bold;
-    ${maxWidth.tablet} {
-      font-size: 1.25rem;
-      line-height: 1.6;
-    };
-  `;
-  const subHeadingBorderTopCss = `
-    border-top: 1px solid ${palette.neutral["86"]};
-    margin-top: 50px;
-  `;
-  const subHeadingCss = `
-    ${subHeadingBorderTopCss}
-    ${subHeadingTitleCss}
-  `;
-
-  return (
-    <FlowStartMultipleProductDetailHandler
-      {...props}
-      headingPrefix={"Manage"}
-      hideHeading
-      hasLeftNav={{
-        pageTitle: "Manage subscription",
-        selectedNavItem: navLinks.accountOverview
-      }}
-      supportRefererSuffix="manage_subscription_flow"
-      loadingMessagePrefix="Retrieving details of your"
-      cancelledExplainer={`This ${props.productType.friendlyName} has been cancelled. Please contact us if you would like to re-start this ${props.productType.friendlyName}, make any amendments or need further help.`}
-      allowCancelledSubscription
-      singleProductDetailRenderer={(
-        routeableStepProps: RouteableStepProps,
-        productDetail: ProductDetail
-      ) => {
-        const mainPlan = getMainPlan(productDetail.subscription);
-
-        const futurePlan = getFuturePlanIfVisible(productDetail.subscription);
-
-        const productType =
-          ProductTypes.subscriptions.mapGroupedToSpecific?.(productDetail) ||
-          props.productType;
-
-        const hasCancellationPending = productDetail.subscription.cancelledAt;
-
-        return (
-          <>
-            <PageHeaderContainer
-              selectedNavItem={navLinks.accountOverview}
-              title="Manage subscription"
-              breadcrumbs={[
-                {
-                  title: navLinks.accountOverview.title,
-                  link: navLinks.accountOverview.link
-                },
-                {
-                  title: "Manage subscription",
-                  currentPage: true
-                }
-              ]}
-            />
-            <PageNavAndContentContainer
-              selectedNavItem={navLinks.accountOverview}
-            >
-              {productDetail.alertText && (
-                <ProblemAlert
-                  title="A payment needs your attention"
-                  message={productDetail.alertText}
-                  button={{
-                    title: "Update payment method",
-                    link: `/payment/${productType.urlPart}`,
-                    state: productDetail
-                  }}
-                  additionalcss={css`
-                    margin-top: 30px;
-                  `}
-                />
-              )}
-              <div
-                css={css`
-                  ${subHeadingBorderTopCss}
-                  display: flex;
-                  align-items: start;
-                  justify-content: space-between;
-                `}
-              >
-                <h2
-                  css={css`
-                    ${subHeadingTitleCss}
-                    margin: 0;
-                  `}
-                >
-                  {productType.productTitle(mainPlan)}
-                </h2>
-                {isGift(productDetail.subscription) && (
-                  <i
-                    css={css`
-                      margin: 4px 0 0 ${space[3]}px;
-                    `}
-                  >
-                    <GiftIcon alignArrowToThisSide={"left"} />
-                  </i>
-                )}
-              </div>
-              {hasCancellationPending && (
-                <p
-                  css={css`
-                    ${textSans.medium()};
-                  `}
-                >
-                  <ErrorIcon fill={palette.brandYellow[200]} />
-                  <span
-                    css={css`
-                      margin-left: ${space[2]}px;
-                    `}
-                  >
-                    Your subscription has been <strong>cancelled</strong>. You
-                    are able to access your subscription until{" "}
-                    <strong>
-                      {formatDateStr(productDetail.subscription.end)}
-                    </strong>
-                  </span>
-                  .
-                </p>
-              )}
-              <ProductDescriptionListTable
-                content={[
-                  {
-                    title: "Subscription ID",
-                    value: productDetail.subscription.subscriptionId
-                  },
-                  {
-                    title: "Start date",
-                    value: productDetail.subscription.start
-                      ? formatDateStr(productDetail.subscription.start)
-                      : "-"
-                  }
-                ]}
-              />
-              <h2
-                css={css`
-                  ${subHeadingCss}
-                `}
-              >
-                Payment
-              </h2>
-              {isSixForSix(mainPlan.name) &&
-                isPaidSubscriptionPlan(mainPlan) &&
-                !hasCancellationPending && (
-                  <p
-                    css={css`
-                      ${textSans.medium()};
-                    `}
-                  >
-                    This subscription is still in the initial '6 issues for{" "}
-                    {mainPlan.currency}6' promotional period.
-                  </p>
-                )}
-              <ProductDescriptionListTable
-                borderColour={palette.neutral[86]}
-                alternateRowBgColors
-                content={[
-                  ...(isPaidSubscriptionPlan(mainPlan) &&
-                  productDetail.subscription.autoRenew &&
-                  !hasCancellationPending
-                    ? [
-                        {
-                          title: `Next ${augmentInterval(
-                            productDetail.subscription.currentPlans.length !==
-                              0 && isSixForSix(mainPlan.name)
-                              ? futurePlan.interval
-                              : mainPlan.interval
-                          )} payment`,
-                          value: `${mainPlan.currency}${(
-                            (productDetail.subscription.nextPaymentPrice ||
-                              mainPlan.amount) / 100.0
-                          ).toFixed(2)}${" "}${mainPlan.currencyISO}`
-                        },
-                        {
-                          title: "Next payment date",
-                          ...(productDetail.subscription.nextPaymentDate && {
-                            value: formatDateStr(
-                              productDetail.subscription.currentPlans.length ===
-                                0
-                                ? mainPlan.start
-                                : productDetail.subscription.nextPaymentDate
-                            )
-                          })
-                        }
-                      ]
-                    : []),
-                  {
-                    title: "Payment method",
-                    value: productDetail.isPaidTier ? (
-                      <>
-                        {productDetail.subscription.card && (
-                          <CardDisplay
-                            margin="0"
-                            inErrorState={!!productDetail.alertText}
-                            {...productDetail.subscription.card}
-                          />
-                        )}
-                        {productDetail.subscription.payPalEmail && (
-                          <PayPalDisplay
-                            payPalId={productDetail.subscription.payPalEmail}
-                            shouldIncludePrefixCopy
-                          />
-                        )}
-                        {productDetail.subscription.mandate && (
-                          <DirectDebitDisplay
-                            inErrorState={!!productDetail.alertText}
-                            {...productDetail.subscription.mandate}
-                          />
-                        )}
-                        {productDetail.subscription
-                          .stripePublicKeyForCardAddition && (
-                          <span>No Payment Method</span>
-                        )}
-                      </>
-                    ) : (
-                      <span>FREE</span>
-                    ),
-                    spanTwoCols: productDetail.subscription.payPalEmail
-                      ? true
-                      : undefined
-                  },
-                  {
-                    title: "Expiry date",
-                    ...(productDetail.subscription.card?.expiry && {
-                      value: `${
-                        productDetail.subscription.card.expiry.month < 10
-                          ? "0"
-                          : ""
-                      }${productDetail.subscription.card.expiry.month}
-                    ${" / "}
-                    ${productDetail.subscription.card.expiry.year}`
-                    })
-                  }
-                ]}
-              />
-              {productDetail.isPaidTier && (
-                <LinkButton
-                  colour={
-                    productDetail.alertText
-                      ? palette.brand[400]
-                      : palette.brand[800]
-                  }
-                  textColour={
-                    productDetail.alertText
-                      ? palette.neutral[100]
-                      : palette.brand[400]
-                  }
-                  fontWeight={"bold"}
-                  alert={!!productDetail.alertText}
-                  text="Update payment method"
-                  to={`/payment/${productType.urlPart}`}
-                  state={productDetail}
-                />
-              )}
-              {productType.delivery?.showAddress?.(
-                productDetail.subscription
-              ) &&
-                productDetail.subscription.deliveryAddress && (
-                  <>
-                    <h2
-                      css={css`
-                        ${subHeadingCss}
-                      `}
-                    >
-                      Delivery address
-                    </h2>
-                    <ProductDescriptionListTable
-                      borderColour={palette.neutral[86]}
-                      content={[
-                        {
-                          title: "Address",
-                          value: (
-                            <DeliveryAddressDisplay
-                              {...productDetail.subscription.deliveryAddress}
-                            />
-                          ),
-                          spanTwoCols: true
-                        }
-                      ]}
-                    />
-                    <LinkButton
-                      colour={palette.brand[800]}
-                      textColour={palette.brand[400]}
-                      fontWeight="bold"
-                      text="Manage delivery address"
-                      to={`/delivery/${productType.urlPart}/address`}
-                      state={productDetail}
-                    />
-                  </>
-                )}
-              {hasDeliveryRecordsFlow(productType) && (
-                <>
-                  <h2
-                    css={css`
-                      ${subHeadingCss}
-                    `}
-                  >
-                    Delivery history
-                  </h2>
-                  <p>Check delivery history and report an issue.</p>
-                  <LinkButton
-                    colour={palette.brand[800]}
-                    textColour={palette.brand[400]}
-                    fontWeight="bold"
-                    text="Manage delivery history"
-                    to={`/delivery/${productType.urlPart}/records`}
-                    state={productDetail}
-                  />
-                </>
-              )}
-              {shouldHaveHolidayStopsFlow(productType) &&
-                productDetail.subscription.autoRenew &&
-                !hasCancellationPending && (
-                  <>
-                    <h2
-                      css={css`
-                        ${subHeadingCss}
-                      `}
-                    >
-                      Going on holiday?
-                    </h2>
-                    <p>
-                      Don’t fret - you can manage your suspensions by clicking
-                      the button below. You will be credited for each suspended
-                      {productType.holidayStops.issueKeyword} on the first bill
-                      after the suspension date.
-                    </p>
-                    <LinkButton
-                      colour={palette.brand[800]}
-                      textColour={palette.brand[400]}
-                      fontWeight="bold"
-                      text="Manage suspensions"
-                      to={`/suspend/${productType.urlPart}`}
-                      state={productDetail}
-                    />
-                  </>
-                )}
-              {!productDetail.subscription.autoRenew &&
-                productType.renewalMetadata && (
-                  <>
-                    <h2
-                      css={css`
-                        ${subHeadingCss}
-                      `}
-                    >
-                      Renewal
-                    </h2>
-                    <p>
-                      To renew your one-off {productType.friendlyName}, please
-                      contact us.
-                    </p>
-                    <CallCentreEmailAndNumbers />
-                    <p>
-                      Alternatively, if you would prefer to start a recurring{" "}
-                      {productType.friendlyName} you can explore payment options
-                      and subscribe online by clicking the button below.
-                    </p>
-                    <SupportTheGuardianButton
-                      {...productType.renewalMetadata}
-                      fontWeight="bold"
-                      textColour={palette.neutral[100]}
-                      colour={palette.brand[400]}
-                      notPrimary
-                    />
-                  </>
-                )}
-              {productType.cancellation && !hasCancellationPending && (
-                <Link
-                  css={css`
-                  display: block;
-                  float: right;
-                  margin: ${space[24]}px 0 0 auto;
-                  ${textSans.medium()}
-                  color: ${palette.brand["500"]};
-                  border-bottom: 1px solid ${palette.neutral["100"]};
-                  transition: border-color .15s ease-out;
-                  :hover: {
-                    borderBottom: 1px solid ${palette.brand["400"]};
-                  }
-                `}
-                  to={"/cancel/" + productType.urlPart}
-                  state={productDetail}
-                >
-                  Cancel subscription
-                </Link>
-              )}
-            </PageNavAndContentContainer>
-          </>
-        );
-      }}
-    />
-  );
-};
